+++ MaidChan.py
@@ -5,7 +5,6 @@ from discord.ext import commands
 
 
 class MaidChan(commands.Bot):
-    __instance = None
     __TOKEN = ""
     __PREFIX = ""
 
@@ -29,21 +28,12 @@ class MaidChan(commands.Bot):
                 self.load_extension(f'extensions.{extension[:-3]}')
         pass
 
-    @staticmethod
-    def instance():
-        if MaidChan.__instance is None:
-            MaidChan()
-        return MaidChan.__instance
-
-    @staticmethod
-    def run():
-        MaidChan.instance()
+    def run(self):
+        super().run(self.TOKEN)
         pass
 
-    @staticmethod
-    def get_token():
-        return MaidChan.__TOKEN
+    def get_token(self):
+        return self.TOKEN
 
-    @staticmethod
-    def get_prefix():
-        return MaidChan.__PREFIX
+    def get_prefix(self):
+        return self.TOKEN
